Use new hashing function for TranslationBlock
Avoid to put all the TB with the same physical address and the same
virtual PC in the same bucket.
Indeed, when a shared library is mapped in a large number of processes,
the code is re-translated several times (once per mapping) but all the
TB end up the same bucket defeating the purpose of the hash table and
producing very slow list search.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>